requirements.txt
psutil==5.8.0
Pillow==8.3.2